+++ src/main/java/com/oms/entity/OrderItem.java
@@ -18,7 +18,7 @@ public class OrderItem {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long orderItemId;
+    private int orderItemId;
 
     private int quantity;
