+++ src/main/java/com/oms/entity/Payment.java
@@ -19,7 +19,7 @@ public class Payment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long paymentId;
+    private int paymentId;
 
     private String paymentMethod;
