+++ src/main/java/com/oms/repository/InventoryRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface InventoryRepository extends JpaRepository<Inventory,Long> {
+public interface InventoryRepository extends JpaRepository<Inventory,Integer> {
 
-    Optional<Inventory> findByProduct_ProductIdAndWarehouse_WarehouseId(Long productId,Long warehouseId);
+    Optional<Inventory> findByProduct_ProductIdAndWarehouse_WarehouseId(int productId,int warehouseId);
 
-    List<Inventory> findByProduct_ProductId(Long productId);
+    List<Inventory> findByProduct_ProductId(int productId);
 
-    List<Inventory> findByWarehouse_WarehouseId(Long warehouseId);
+    List<Inventory> findByWarehouse_WarehouseId(int warehouseId);
 }
