+++ src/main/java/com/oms/repository/OrderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Orders, Long> {
+public interface OrderRepository extends JpaRepository<Orders, Integer> {
 
     List<Orders> findByCustomer(Customer customerId);
     //Alternative to above method is
