+++ src/main/java/com/oms/repository/PaymentRepository.java
@@ -7,10 +7,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 
     //Fetch Payments for a specific order
-    List<Payment> findByOrder_OrderId(Long orderId);
+    List<Payment> findByOrder_OrderId(int orderId);
 
     //Fetch by payment status(SUCCESS,FAILED)
     List<Payment> findByPaymentStatus(String paymentStatus);
