+++ src/main/java/com/oms/repository/WarehouseRepository.java
@@ -3,5 +3,5 @@ package com.oms.repository;
 import com.oms.entity.Warehouse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WarehouseRepository extends JpaRepository<Warehouse, Long> {
+public interface WarehouseRepository extends JpaRepository<Warehouse, Integer> {
 }
